WIP/DEV:Rewriting the client side following mdservice major refactor
