+++ metadataclient/api.py
@@ -1,7 +1,7 @@
 # Data retrieval
-from metadataclient.commands import (find_events, find_descriptors, find_run_starts,
-                                     find_last, find_run_stops, server_connect)
+# from metadataclient.commands import (find_events, find_descriptors, find_run_starts,
+#                                      find_last, find_run_stops, server_connect)
 # Data insertion
-from metadataclient.commands import (insert_event, insert_descriptor, insert_run_start,
-                                     insert_run_stop, bulk_insert_events)
-from doct import Document
+# from metadataclient.commands import (insert_event, insert_descriptor, insert_run_start,
+#                                      insert_run_stop, bulk_insert_events)
+# from doct import Document
